+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## next
 
 - Use target Kotlin API Version 1.3 [#750](../../issues/750)
+- Migrate `SignPluginTask` to use the Marketplace ZIP Signer CLI
 
 ## 1.1.4